[NAE-1546] Plugin library
- plugin registration server
